Sentences assigned by ENPOL
Data that reports and captures the length of different sentences assigned to the surveyed prison population. 
For each case (row in the table), the length of time assigned to prison is reported in the variable SENTENCE, reported in years. For one case with a sentence of 0.25, for example, means that they were assigned one quarter of a year (or three months) in jail.
The data is from INEGI:
https://en.www.inegi.org.mx/programas/enpol/2021/